Make warning

+++ reactivedb_rust_client/src/lib.rs
@@ -7,13 +7,14 @@ use crate::types::EntryValue;
     use crate::types::EntryBuilder;
     use crate::types::DBRequest;
     use crate::client::Client;
+    // DO NOT RUN THESE TESTS
     #[test]
     fn test_insert() {
         let mut client = Client::new("127.0.0.1:1108");
         client.open_connection().unwrap();
         let mut entry_builder = EntryBuilder::new();
-        entry_builder.column("testForIndex", EntryValue::Integer(8));
-        entry_builder.column("testForIteration", EntryValue::Integer(18));
+        entry_builder.column("testForIndex", EntryValue::Integer(11));
+        entry_builder.column("testForIteration", EntryValue::Integer(11));
 
         let insert_request = DBRequest::new_insert("testTable".to_string(), entry_builder.build());
         print!("\n{:?}\n", insert_request);
